added the public zip containing test files
The file contains all the test files.
Unzip it, navigate to index html file and double click to open it.
once opened, fill the form and click the button. 
The form will submit, show the post/comment in the Latest Comment field and clear the rest of the form ready for next inputs.
Once, another input is entered. The existing text in the Latest Comment section, will disappear into a hidden field, which can be viewed via console.